Add timeout when readPin fails to read setPin
Often you can experience a short between the set pin and the read pin. This can happen if the resistor fails for example or it is poorly soldered. This commit adds a timeout after 1000 readings (best guess for a good metric, roughly based on propogation should happen after x micro seconds, so we will time out after x milliseconds, which should still be imperceptable to a human).

Adding the timeout will, for example, allow the touch light to still talk to the cloud even if the read pin fails to read.

I tested 1000 and it appeared to make the light unresponsive so I bump the timeout up to 10000 reads before timing out.