Use distance metric (abs value) for signal metrics
In the case of SENSITIVITY, we want to consider
measurements that fall outside an expected range.
This indicates a valid "touch" signal. I am not
clear on if the polarity (looking for an increase
in delay is expected), however, I have encountered
real life scenarios where the delay drops when the
light is touched. Hence the change to overall
distance metric as opposed to only an increase.

In the case of BASELINE_SENSITIVITY, we are looking
for metrics that fall outside an acceptable deviation.
These measurements do not qualify for the moving average,
as they are outliers.