Reduce PERIODIC_UPDATE_TIME to 60s
Particle has started to put limits on number of messages
sent per month. PERIODIC_UPDATE_TIME is a nice-to-have
in that the first message sent on touch should update
the rest of the touch lights.

The periodic broadcast is a safety mechanism to ensure
that the lights become synchronized in a bound period
of time. Reducing the update time to 60 seconds ensures
this synchronization happens after 60 seconds in the
worst case. In practice, lights should be synchronized
instantaneously regardless of this update time.